feat: unwrap question namespace to flat exports + barrel
